+++ arbitragelab/pairs_selection/__init__.py
@@ -1,0 +1,5 @@
+"""
+Class which implements various functions used to filter-out cointegrated pairs.
+"""
+
+from arbitragelab.pairs_selection.cointegration import CointegrationPairsSelector