+++ docs/source/changelog.rst
@@ -12,9 +12,21 @@ Changelog
 ..
     For Help: https://releases.readthedocs.io/en/latest/index.html
 
+* :bug:`52` Fixed issue with too many function calls in web analytics.
+* :feature:`48` ML Approach Pairs Selection Module made more flexible - clustering and selection steps are now separate.
+* :support:`48` ML Approach Pairs Selection Module documentation updated.
+* :feature:`48` Hedge Ratio Estimation Module added with OLS, TLS, and Minimum HL Methods.
+* :support:`48` Hedge Ratio Estimation Module documentation.
+* :bug:`48` Fixed bug in ML Approach Pairs Selector hedge ratio calculation (previously had included intercept).
+* :feature:`45` Pearson Strategy added to the Distance Approach Module.
+* :support:`45` Pearson Strategy documentation.
+* :feature:`46` Optimal Convergence Model added to the Stochastic Control Approach Module.
+* :support:`46` Optimal Convergence Model documentation.
+* :feature:`49` Cointegration and OU Model Tear Sheets added to the Visualization Module.
+* :support:`49` Cointegration and OU Model Tear Sheets documentation.
 * :support:`50` Updated documentation theme to hudsonthames-sphinx-docs.
 
-* :release:`0.4.10 <2021-04-15>`
+* :release:`0.4.1 <2021-04-15>`
 * :feature:`43` OU Model Jurek and OU Model Mudchanatongsuk added to the Stochastic Control Approach Module.
 * :support:`43` OU Model Jurek and OU Model Mudchanatongsuk documentation.
 * :feature:`44` CVine Copula and CVine Copula Strategy added to the Copula Approach Module.